Cisco CML YAML INE EIGRP TOPOLOGY lecture
If you upload this file on Cisco CML it will save you the trouble to create the topology, and it will assit you in following along with INE's lecture. 

To upload the file, go to your CML Dashboard, and click Import on the top right. Then Select the YAML file and stanr the lab.